fix per reviews
