+++ Sources/Optimizely+Decide/OptimizelyClient+Decide.swift
@@ -30,5 +30,145 @@ extension OptimizelyClient {
                                   attributes: [String: Any]? = nil) -> OptimizelyUserContext {
         return OptimizelyUserContext(optimizely: self, userId: userId, attributes: attributes)
     }
+    
+    func decide(user: OptimizelyUserContext,
+                key: String,
+                options: [OptimizelyDecideOption]? = nil) -> OptimizelyDecision {
         
+        guard let config = self.config else {
+            return OptimizelyDecision.errorDecision(key: key, user: user, error: .sdkNotReady)
+        }
+        
+        guard let feature = config.getFeatureFlag(key: key) else {
+            return OptimizelyDecision.errorDecision(key: key, user: user, error: .featureKeyInvalid(key))
+        }
+        
+        let userId = user.userId
+        let attributes = user.attributes
+        let allOptions = defaultDecideOptions + (options ?? [])
+        let decisionReasons = DecisionReasons()
+        var sentEvent = false
+        var enabled = false
+    
+        let decision = decisionService.getVariationForFeature(config: config,
+                                                              featureFlag: feature,
+                                                              userId: userId,
+                                                              attributes: attributes,
+                                                              options: allOptions)
+        
+        if let featureEnabled = decision?.variation?.featureEnabled {
+            enabled = featureEnabled
+        }
+        
+        var variableMap = [String : Any]()
+        if !allOptions.contains(.excludeVariables) {
+            variableMap = getDecisionVariableMap(feature: feature,
+                                                 variation: decision?.variation,
+                                                 enabled: enabled,
+                                                 reasons: decisionReasons)
+        }
+        
+        var optimizelyJSON: OptimizelyJSON
+        if let opt = OptimizelyJSON(map: variableMap) {
+            optimizelyJSON = opt
+        } else {
+            decisionReasons.addError(OptimizelyError.invalidJSONVariable)
+            optimizelyJSON = OptimizelyJSON.createEmpty()
+        }
+
+        let reasonsToReport = decisionReasons.getReasonsToReport(options: allOptions)
+        
+        if let experimentDecision = decision?.experiment, let variationDecision = decision?.variation {
+            if !allOptions.contains(.disableDecisionEvent) {
+                sendImpressionEvent(experiment: experimentDecision,
+                                    variation: variationDecision,
+                                    userId: userId,
+                                    attributes: attributes)
+                sentEvent = true
+            }
+        }
+        
+        // TODO: add ruleKey values when available later. Use a copy of experimentKey for now.
+        let ruleKey = decision?.experiment?.key
+        
+        sendDecisionNotification(userId: userId,
+                                 attributes: attributes,
+                                 decisionInfo: DecisionInfo(decisionType: .flag,
+                                                            experiment: decision?.experiment,
+                                                            variation: decision?.variation,
+                                                            feature: feature,
+                                                            featureEnabled: enabled,
+                                                            variableValues: variableMap,
+                                                            ruleKey: ruleKey,
+                                                            reasons: reasonsToReport,
+                                                            sentEvent: sentEvent))
+        
+        return OptimizelyDecision(variationKey: decision?.variation?.key,
+                                  enabled: enabled,
+                                  variables: optimizelyJSON,
+                                  ruleKey: ruleKey,
+                                  flagKey: feature.key,
+                                  userContext: user,
+                                  reasons: reasonsToReport)
+    }
+    
+    func decide(user: OptimizelyUserContext, keys: [String], options: [OptimizelyDecideOption]? = nil) -> [String: OptimizelyDecision] {
+        return [:]
+    }
+
+    func decideAll(user: OptimizelyUserContext, options: [OptimizelyDecideOption]? = nil) -> [String: OptimizelyDecision] {
+        return [:]
+    }
+
+}
+
+// MARK: - Utils
+
+extension OptimizelyClient {
+
+    func getDecisionVariableMap(feature: FeatureFlag,
+                                variation: Variation?,
+                                enabled: Bool,
+                                reasons: DecisionReasons) -> [String: Any] {
+        var variableMap = [String: Any]()
+        
+        for (_, v) in feature.variablesMap {
+            var featureValue = v.value
+            if enabled, let variable = variation?.getVariable(id: v.id) {
+                featureValue = variable.value
+            }
+                                    
+            if let value = parseFeatureVaraible(value: featureValue, type: v.type) {
+                variableMap[v.key] = value
+            } else {
+                let info = OptimizelyError.variableValueInvalid(v.key)
+                logger.e(info)
+                reasons.addError(info)
+            }
+        }
+        
+        return variableMap
+    }
+    
+    func parseFeatureVaraible(value: String, type: String) -> Any? {
+        var valueParsed: Any? = value
+        
+        if let valueType = Constants.VariableValueType(rawValue: type) {
+            switch valueType {
+            case .string:
+                break
+            case .integer:
+                valueParsed = Int(value)
+            case .double:
+                valueParsed = Double(value)
+            case .boolean:
+                valueParsed = Bool(value)
+            case .json:
+                valueParsed = OptimizelyJSON(payload: value)?.toMap()
+            }
+        }
+        
+        return valueParsed
+    }
+    
 }
